refactor: robot view handling -> Robot.kt
