+++ android/app/src/main/java/com/github/halspals/smarthomeadapters/smarthomeadapters/model/Robot.kt
@@ -1,7 +1,23 @@
 package com.github.halspals.smarthomeadapters.smarthomeadapters.model
 
+import android.content.Context
+import android.util.Log
+import android.view.MotionEvent
+import android.widget.ImageView
+import android.widget.TextView
+import com.github.halspals.smarthomeadapters.smarthomeadapters.EditRobotFragment
+import com.github.halspals.smarthomeadapters.smarthomeadapters.MainActivity
 import com.github.halspals.smarthomeadapters.smarthomeadapters.R
+import com.github.halspals.smarthomeadapters.smarthomeadapters.RestApiService
 import com.google.gson.annotations.SerializedName
+import kotlinx.android.synthetic.main.activity_main.*
+import net.openid.appauth.AuthorizationException
+import okhttp3.ResponseBody
+import org.jetbrains.anko.design.snackbar
+import org.jetbrains.anko.toast
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
 
 data class RobotStatus(
         @SerializedName("value")
@@ -13,6 +29,14 @@ data class RobotStatus(
         @SerializedName("max")
         val max: Int
 )
+
+/**
+ * Model for the body of a robot registration call
+ *
+ * @property nickname the nickname to set for the robot
+ * @property robotType the type of functionality the robot provides
+ */
+data class RobotRegistrationBody(val nickname: String, val robotType: String)
 
 /**
  * Model for a smart home adapter robot
@@ -39,12 +63,201 @@ data class Robot(
         val ADD_ROBOT = Robot("","","","",
                 RobotStatus(false,0,0,0))
     }
-}
 
-/**
- * Model for the body of a robot registration call
- *
- * @property nickname the nickname to set for the robot
- * @property robotType the type of functionality the robot provides
- */
-data class RobotRegistrationBody(val nickname: String, val robotType: String)
+    private val tag = "RobotClass"
+
+    internal fun updateViews(
+            context: Context,
+            robotCircle: ImageView? = null,
+            robotIcon: ImageView? = null,
+            robotRangeText: TextView? = null
+    ) {
+        when (robotInterfaceType) {
+            Robot.INTERFACE_TYPE_TOGGLE -> {
+                robotCircle?.setColorFilter(
+                        if (robotStatus.value) {
+                            context.getColor(R.color.colorToggleOn)
+                        } else {
+                            context.getColor(R.color.colorToggleOff)
+                        }
+                )
+            }
+
+            Robot.INTERFACE_TYPE_RANGE -> {
+                robotRangeText?.text = robotStatus.current.toString()
+            }
+
+            else -> TODO("NO OTHER ROBOT INTERFACE TYPE EXPECTED")
+        }
+
+        when (robotType) {
+            Robot.ROBOT_TYPE_SWITCH -> {
+                robotIcon?.setImageResource(
+                        if (robotStatus.value) {
+                            R.drawable.ic_light_on
+                        } else {
+                            R.drawable.ic_light_off
+                        }
+                )
+            }
+
+            Robot.ROBOT_TYPE_THERMOSTAT -> {
+                // Display no image
+            }
+
+            Robot.ROBOT_TYPE_BOLTLOCK -> {
+                robotIcon?.setImageResource(
+                        if (robotStatus.value) {
+                            R.drawable.basic_lock
+                        } else {
+                            R.drawable.basic_lock_open
+                        }
+                )
+            }
+
+            else -> TODO("NO OTHER ROBOT TYPE EXPECTED")
+
+        }
+    }
+
+    @SuppressWarnings("ClickableViewAccessibility")
+    internal fun setViewEvents(
+            parent: MainActivity,
+            robotCircle: ImageView? = null,
+            robotIcon: ImageView? = null,
+            robotRangeText: TextView? = null
+    ) {
+        // configure interactions with the robot
+        when (robotInterfaceType) {
+            Robot.INTERFACE_TYPE_TOGGLE -> {
+                robotCircle?.setOnClickListener { _ ->
+                    if (parent.isInEditMode) {
+                        parent.robotToEdit = this
+                        parent.startFragment(EditRobotFragment())
+                    } else {
+                        onToggle(parent, robotCircle, robotIcon)
+                    }
+                }
+            }
+
+            Robot.INTERFACE_TYPE_RANGE -> {
+                robotCircle?.setOnTouchListener { _, motionEvent ->
+                    if (parent.isInEditMode) {
+                        parent.robotToEdit = this
+                        parent.startFragment(EditRobotFragment())
+                    } else {
+                        if (motionEvent.action == MotionEvent.ACTION_UP) {
+                            onSeek(parent)
+                        } else {
+                            if (motionEvent.y > robotCircle.y) {
+                                robotStatus.current++
+                                Log.d(tag, "Increasing current value of $this")
+                            } else {
+                                robotStatus.current--
+                            }
+                            Log.d(tag, "Decreasing current calue of $this")
+                            updateViews(context = parent, robotRangeText = robotRangeText)
+                        }
+                    }
+                    true
+                }
+            }
+        }
+    }
+
+    /**
+     * Handles a toggle-type event for a robot, updating its value and sending it to the server.
+     */
+    private fun onToggle(parent: MainActivity, robotCircle: ImageView?, robotIcon: ImageView?) {
+        Log.wtf(tag, "[onToggle]: robot is $this)")
+
+        // Send the update to the server
+        parent.authState.performActionWithFreshTokens(parent.authService)
+        { accessToken: String?, _: String?, ex: AuthorizationException? ->
+            if (accessToken == null) {
+                Log.e(tag, "[onSwitch] got null access token, exception: $ex")
+            } else {
+                parent.restApiService
+                        .robotToggle(id, !robotStatus.value, accessToken, mapOf())
+                        .enqueue(object: Callback<ResponseBody> {
+
+                            override fun onResponse(
+                                    call: Call<ResponseBody>,
+                                    response: Response<ResponseBody>) {
+
+                                if (response.isSuccessful) {
+                                    parent.toast("Success")
+                                    Log.d(tag, "[onToggle] Server accepted setting" +
+                                            "toggle to ${!robotStatus.value}")
+                                    robotStatus.value = !robotStatus.value
+                                    updateViews(
+                                            context = parent,
+                                            robotCircle = robotCircle,
+                                            robotIcon = robotIcon)
+
+                                } else {
+                                    val error = RestApiService.extractErrorFromResponse(response)
+                                    Log.e(tag, "[onToggle] Unsuccessful, "
+                                            + "error: $error")
+                                    if (error != null) {
+                                        parent.snackbar_layout.snackbar(error)
+                                    }
+                                }
+                            }
+
+                            override fun onFailure(call: Call<ResponseBody>, t: Throwable) {
+                                val error = t.message
+                                Log.e(tag, "[onToggle] FAILED, error: $error")
+                                if (error != null) {
+                                    parent.snackbar_layout.snackbar(error)
+                                }
+                            }
+                        })
+            }
+        }
+    }
+
+    /**
+     * onSeek is called whenever a range-type robot wants to make an API call to change state
+     */
+    private fun onSeek(parent: MainActivity) {
+        Log.d(tag, "onSeek(${robotStatus.current})")
+
+        parent.authState.performActionWithFreshTokens(parent.authService)
+        { accessToken: String?, _: String?, ex: AuthorizationException? ->
+            if (accessToken == null) {
+                Log.e(tag, "[onSeek] got null access token, exception: $ex")
+            } else {
+                parent.restApiService
+                        .robotRange(id, robotStatus.current, accessToken, mapOf())
+                        .enqueue(object : Callback<ResponseBody> {
+
+                            override fun onResponse(
+                                    call: Call<ResponseBody>,
+                                    response: Response<ResponseBody>) {
+
+                                if (response.isSuccessful) {
+                                    parent.toast("Success")
+                                    Log.d(tag, "Server accepted setting range to ${robotStatus.current}")
+                                } else {
+                                    val error = RestApiService.extractErrorFromResponse(response)
+                                    Log.e(tag, "Setting the range was unsuccessful, "
+                                            + "error: $error")
+                                    if (error != null) {
+                                        parent.snackbar_layout.snackbar(error)
+                                    }
+                                }
+                            }
+
+                            override fun onFailure(call: Call<ResponseBody>, t: Throwable) {
+                                val error = t.message
+                                Log.e(tag, "onSeek(${robotStatus.current}) FAILED, error: $error")
+                                if (error != null) {
+                                    parent.snackbar_layout.snackbar(error)
+                                }
+                            }
+                        })
+            }
+        }
+    }
+}
